fixed adc

+++ main/include/mics4514_if.h
@@ -9,7 +9,7 @@
 #define MAIN_INCLUDE_MICS4514_IF_H_
 
 void MICS4514_Initialize(void);
-void MICS4514_Poll(uint16_t *ox_val, uint16_t *red_val);
+void MICS4514_Poll(int *ox_val, int *red_val);
 
 
 #endif /* MAIN_INCLUDE_MICS4514_IF_H_ */
